Add: traveler

+++ dp/fib/memo.cpp
@@ -1,7 +1,7 @@
 #include <bits/stdc++.h>
 using namespace std;
 
-int fibonacci(int n, vector<int> &memo) {
+long long fibonacci(long long n, vector<long long> &memo) {
   if (memo[n] != -1)
     return memo[n];
   if (n <= 2)
@@ -15,7 +15,7 @@ int main() {
   int n;
   cin >> n;
 
-  vector<int> memo(100000000, -1);
+  vector<long long> memo(100000000, -1);
   cout << fibonacci(n, memo) << endl;
   return 0;
 }
